fast export
exports the library in a swift app instead of using applescript. this cuts the full export latency down from ~40 minutes to ~2.5 minutes (on a ~9500 track library). it also adds a fast export mode that re-uses existing track & artwork file md5s to cut the export time down to ~6 seconds. while it is slightly more involved to do the first export, i think the latency savings are worth it.

other misc changes:
* drop tables instead of deleting the entire database (can reduce permissions on the database user)
* run multiple inserts in one query (tbh this probably accounts for a large portion of the speedup)
* add database_password option
* clean up config objects in ruby
* move the update library portion to a separate module